change the input text field into a text view
In this case a text view makes a lot more sense. I probably avoided it the first place so I didn't have to deal with the auto resizing.
Before, you had to press the enter button on the keyboard to add a new note. Now you press the submit button next to the text view.